refactor: use UI components
