+++ frontend/src/components/forms/Login/LoginForm/LoginForm.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/index.scss';
+@import '../../../../styles/index.scss';
 .form {
   display: flex;
   flex-direction: column;
