+++ frontend/src/pages/signin/index.jsx
@@ -1,4 +1,4 @@
-import LoginForm from '../../components/forms/Login/LoginForm';
+import LoginForm from '../../components/forms/Login/LoginForm/LoginForm';
 
 export default function SignInPage() {
   return <LoginForm />;
